% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{discent_sim_dat}
\alias{discent_sim_dat}
\title{Malaria Gene Gene Flow up a Mountain}
\format{
A list with the following items:
\describe{
\item{locatdat}{Dataframe: Mock latitudes and longitudes for each deme that was simualted on a 30x30 square}
\item{gridmig}{Dataframe: The migration matrix that defines gene flow between demes}
\item{migplot}{ggplot: A rasterized plot of the migration matrix}
\item{discent_dat}{Dataframe: The discent data in the correct format}
}
}
\usage{
discent_sim_dat
}
\description{
A dataset that was simulated using the \code{https://github.com/nickbrazeau/polySimIBD} package:
\url{https://github.com/nickbrazeau/polySimIBD}, which is a "cartoon-ish" version for malaria
transmission that relies on a discrete-time discrete-loci structured Wright-Fisher model
}
\keyword{datasets}
