% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{drawDistMat}
\alias{drawDistMat}
\title{Simple function for simulating a distance matrix}
\usage{
drawDistMat(nDemes = 5, rateDist = 0.001)
}
\arguments{
\item{nDemes}{integer; number of demes to consider}

\item{rateDist}{numeric; rate in exponential distribution (for drawing distances)}
}
\value{
Distance matrix of class \code{dist}
}
\description{
Simple function for simulating a distance matrix
}
\details{
Draws distances from a squared-normal distribution
}
