% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expand_pairwise}
\alias{expand_pairwise}
\title{Simple function for expanding a pairwise matrix}
\usage{
expand_pairwise(discdf)
}
\arguments{
\item{discdf}{dataframe; that is specific to the DISCent framework. Must have colnames of: "smpl1", "smpl2", "deme1", "deme2", "gendist", "geodist" in correct order}
}
\description{
Simple function for expanding a pairwise matrix
}
\details{
This is an internal function that largely lacks generalizability. Do not recommend external use.
}
