% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBD_simulation_data.R
\docType{data}
\name{IBD_simulation_data}
\alias{IBD_simulation_data}
\title{Simulated Identity by Descent from Isolation by Distance}
\format{
A dataframe with 45 rows and 6 columns:
\describe{
\item{smpl1, smpl2}{Placeholder sample names}
\item{deme1, deme2}{Placeholder discrete demes}
\item{gendist}{Simulated genetic distances based on identity by descent}
\item{geodist}{Simulated geographic distances}
}
}
\source{
A toy dataset generated by basic simulation assuming an exponential relationship
between relatedness and geographic distance. Data is not representative or
generalizable but is simply meant to be used as input for various tests and
function explanations
}
\usage{
IBD_simulation_data
}
\description{
Simulated Identity by Descent from Isolation by Distance
}
\keyword{datasets}
