% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.DISCresult}
\alias{is.DISCresult}
\title{Check if DISCresult S3 Class}
\usage{
is.DISCresult(x)
}
\arguments{
\item{x}{DISC result from deme_inbreeding_spcoef function}
}
\description{
Overload is: function for determining if object is of class DISCresult
}
