% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.vanillaDISCresult}
\alias{print.vanillaDISCresult}
\title{print DISCresult S3 Class}
\usage{
\method{print}{vanillaDISCresult}(x, ...)
}
\arguments{
\item{x}{DISC result from deme_inbreeding_spcoef function}

\item{...}{further arguments passed to or from other methods.}
}
\description{
overload print() function to print summary only
}
