% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_simulator.R
\name{rbetabinom}
\alias{rbetabinom}
\title{Draw from Beta-binomial distribution}
\usage{
rbetabinom(n = 1, k = 10, alpha = 1, beta = 1)
}
\arguments{
\item{n}{number of draws.}

\item{k}{number of binomial trials.}

\item{alpha}{first shape parameter of beta distribution.}

\item{beta}{second shape parameter of beta distribution.}
}
\description{
Draw from Beta-binomial distribution.
}
