% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim_IBDIBD}
\alias{sim_IBDIBD}
\title{Simulate pairwise Identity by Descent accounting for Isolation by Distance}
\usage{
sim_IBDIBD(demesize, distmat, rate, Ft)
}
\arguments{
\item{demesize}{numeric vector; Contains the number of individuals residing within
the respective deme.}

\item{distmat}{numeric matrix; A distance matrix corresponding to the distances
between demes. The \eqn{m x m} matrix should be of the same length as the number of
demes.}

\item{rate}{numeric; Expected rate of decay for isolation by distance.
In effect, a scalar for the distance matrix.}

\item{Ft}{numeric; Inbreeding coefficient parameter in the population
(\emph{i.e.} the total population, \eqn{N_e}).}
}
\description{
\deqn{ \frac{1}{N_i + N_j}(1 - P(D \leq d)) + (1 - \frac{1}{N_i + N_j})(1 - P(D \leq d)) }
}
