% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.DISCresult}
\alias{summary.DISCresult}
\title{Summary of DISCresult S3 Class}
\usage{
\method{summary}{DISCresult}(object, ...)
}
\arguments{
\item{object}{DISCresult Simulation}

\item{...}{further arguments passed to or from other methods.}
}
\description{
overload summary() function.
}
