% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tidyout}
\alias{tidyout}
\title{Tidy Output Generic Method}
\usage{
tidyout(x)
}
\arguments{
\item{x}{Object to tidy (typically a DISCresult)}
}
\value{
Method-specific tidied output
}
\description{
Generic method for tidying output from DISC analysis
}
