% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tidyout.vanillaDISCresult}
\alias{tidyout.vanillaDISCresult}
\title{Tidy Out Sim}
\usage{
\method{tidyout}{vanillaDISCresult}(x)
}
\arguments{
\item{x}{DISC result from deme_inbreeding_spcoef function}
}
\description{
Function for taking output of SIR NE and lifting it over
}
