% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_vanilla.R
\name{wrangle_discentdat}
\alias{wrangle_discentdat}
\title{Wrangle DISC data for efficient input into Cpp}
\usage{
wrangle_discentdat(discdat, normalize_geodist, start_params, locats)
}
\arguments{
\item{discdat}{dataframe; The genetic-geographic data by deme (K)}

\item{normalize_geodist}{boolean; whether geographic distances between demes should be normalized (i.e. Min-Max Feature Scaling: \eqn{X' = \frac{X - X_{min}}{X_{max} - X_{min}} }, which places the geodistances on the scale to \eqn{[0-1]}). Helps increase model stability at the expense of complicating the interpretation of the migration rate parameter.}

\item{start_params}{named double vector; vector of start parameters.}
}
\description{
Internal wrangling function that expands the DISC genetic data into an array and the geographic
distances into a matrix for efficient input into Cpp.
}
