% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hessian.R
\name{calculate_hessian_eigen}
\alias{calculate_hessian_eigen}
\title{Calculate Hessian Matrix from Loss Function}
\usage{
calculate_hessian_eigen(mod, discdat, lambda)
}
\arguments{
\item{mod}{class of DISCresult; output from \code{disc}}

\item{discdat}{dataframe; The genetic-geographic data by deme (K). Must contain columns:
\code{smpl1}, \code{smpl2}, \code{deme1}, \code{deme2}, \code{gendist}, \code{geodist}}

\item{lambda}{double; A quadratic L2 regularization parameter on "m" parameter: \eqn{\lambda m^2}. Default: 0.1}
}
\value{
A list containing:
\itemize{
\item \code{Hessian}: Hessian matrix
\item \code{Eigen}: Eigenvalues and eigenvectors from the Hessian
\item \code{KappaH}: Conditional number from Hessian matrix
}
}
\description{
Calculate the Hessian, eigenvalues/vecotrs, and conditional number from the Loss Function using central finite differences
}
