% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{deme_inbreeding_spcoef}
\alias{deme_inbreeding_spcoef}
\title{Identify Deme Inbreeding Spatial Coefficients in Continuous Space}
\usage{
deme_inbreeding_spcoef(
  K_gendist_geodist,
  start_params = c(),
  m_lowerbound = 0,
  m_upperbound = 1,
  f_learningrate = 1e-04,
  m_learningrate = 1e-10,
  full_matrix = FALSE,
  steps = 1000,
  report_progress = TRUE
)
}
\arguments{
\item{K_gendist_geodist}{dataframe; The genetic-geographic data by deme (K)}

\item{start_params}{named numeric vector; vector of start parameters.}

\item{m_lowerbound}{numeric; a lower bound for the m parameter}

\item{m_upperbound}{numeric; an upper bound for the m parameter}

\item{f_learningrate}{numeric; alpha parameter for how much each inbreeding "step" is weighted in the gradient descent}

\item{m_learningrate}{numeric; alpha parameter for how much each migration "step" is weighted in the gradient descent}

\item{full_matrix}{boolean; whether or not the user entered in a full matrix. If this is a symmetric distance matrix,
this should be set to false If this is an asymetric matrix, this should be set to true In the latter instance,
users will need to have already expanded every IJ and JI combination (which are by default not equivalent)}

\item{steps}{numeric; the number of "steps" as we move down the gradient}

\item{report_progress}{boolean; whether or not a progress bar should be shown as you iterate through steps}
}
\description{
The purpose of this statistic is to identify an inbreeding coefficient, or degree of
relatedness, for a given location in space. We assume that locations in spaces can be
represented as "demes," such that multiple individuals live in the same deme
(i.e. samples are sourced from the same location). The expected pairwise relationship
between two individuals, or samples, is dependent on the each sample's deme's inbreeding
coefficient and the geographic distance between the demes.
}
\details{
The gen.geo.dist dataframe must be named with the following columns:
"smpl1"; "smpl2"; "locat1"; "locat2"; "gendist"; "geodist"; which corresponds to:
Sample 1 Name; Sample 2 Name; Sample 1 Location; Sample 2 Location;
Pairwise Genetic Distance; Pairwise Geographpic Distance. Note, the order of the
columns do not matter but the names of the columns must match.

The start_params vector names must match the cluster names (i.e. clusters must be
have a name that we can match on for the starting relatedness paramerts). In addition,
you must provide a start parameter for "m".
}
