% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_pso.R
\name{deme_inbreeding_spcoef_pso}
\alias{deme_inbreeding_spcoef_pso}
\title{Identify Deme Inbreeding Spatial Coefficients in Continuous Space with
Particle Swarm Meta-Optimization}
\usage{
deme_inbreeding_spcoef_pso(
  discdat,
  m_lowerbound = 1e-10,
  m_upperbound = Inf,
  fi_lowerbound = 0.001,
  fi_upperbound = 0.3,
  flearn_lowerbound = 1e-10,
  flearn_upperbound = 0.01,
  mlearn_lowerbound = 1e-15,
  mlearn_upperbound = 1e-08,
  c1 = 0.1,
  c2 = 0.1,
  w = 0.25,
  b1 = 0.9,
  b2 = 0.999,
  e = 1e-08,
  steps = 1000,
  searchsteps = 100,
  swarmsteps = 50,
  swarmsize = 25,
  thin = 1,
  normalize_geodist = TRUE,
  report_progress = TRUE,
  return_verbose = FALSE
)
}
\arguments{
\item{discdat}{dataframe; The genetic-geographic data by deme (K)}

\item{m_lowerbound}{double; lower limit value for the global "m" parameter; will use a reflected normal within the gradient descent algorithm to adjust any aberrant values}

\item{m_upperbound}{double; upper limit value for the global "m" parameter; will use a reflected normal within the gradient descent algorithm to adjust any aberrant values}

\item{fi_lowerbound}{}

\item{fi_upperbound}{}

\item{flearn_lowerbound}{}

\item{flearn_upperbound}{}

\item{mlearn_lowerbound}{}

\item{mlearn_upperbound}{}

\item{c1}{}

\item{c2}{}

\item{w}{}

\item{b1}{double; Exponential decay rates for the first moment estimate}

\item{b2}{double; Exponential decay rates for the second moment estimate}

\item{e}{double; Epsilon (error) for stability in the Adam optimization algorithm}

\item{steps}{integer; the number of "steps" as we move down the gradient}

\item{searchsteps}{}

\item{swarmsteps}{}

\item{swarmsize}{}

\item{thin}{integer; the number of "steps" to keep as part of the output (i.e. if the user specifies 10, every 10th iteration will be kept)}

\item{normalize_geodist}{boolean; whether geographic distances between demes should be normalized (i.e. Min-Max Feature Scaling: \eqn{X' = \frac{X - X_{min}}{X_{max} - X_{min}} }, which places the geodistances on the scale to \eqn{[0-1]}). Helps increase model stability at the expense of complicating the interpretation of the migration rate parameter.}

\item{report_progress}{boolean; whether or not a progress bar should be shown as you iterate through steps}

\item{return_verbose}{boolean; whether the inbreeding coefficients and migration rate should be returned for every iteration or
only for the final iteration. User will typically not want to store every iteration, which can be memory intensive}
}
\description{
Identify Deme Inbreeding Spatial Coefficients in Continuous Space with
Particle Swarm Meta-Optimization
}
\details{
The gen.geo.dist dataframe must be named with the following columns:
"smpl1"; "smpl2"; "deme1"; "deme2"; "gendist"; "geodist"; which corresponds to:
Sample 1 Name; Sample 2 Name; Sample 1 Location; Sample 2 Location;
Pairwise Genetic Distance; Pairwise Geographpic Distance. Note, the order of the
columns do not matter but the names of the columns must match.

We have implemented coding decisions to not allow the "f" inbreeding coefficients to be negative by using a
logit transformation internally in the code.

Gradient descent is performed using the Adam (adaptive moment estimation) optimization approach. Default values
for moment decay rates, epsilon, and learning rates are taken from \cite{DP Kingma, 2014}.

The Particle Swarm Optimization is a meta-optimization (meta-heuristic) approach that attempts to find optimal
start parameters for the user to avoid a grid-search approach as would be best practices for fine-tuning the gradient descent
approach.
}
