% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{find_grad_params}
\alias{find_grad_params}
\title{Find Optimal Start Gradient Parameters using Grid Search}
\usage{
find_grad_params(
  discdat,
  tempstart_params = c(),
  momentum = 0.9,
  discsteps = 1000,
  downsample = 1000,
  fstartmin = 0.1,
  fstartmax = 0.9,
  fstartby = 100,
  mstartmin = 1e-15,
  mstartmax = 0.001,
  mstartby = 100,
  flearnmin = 1e-15,
  flearnmax = 0.01,
  flearnby = 100,
  mlearnmin = 1e-15,
  mlearnmax = 0.01,
  mlearnby = 100
)
}
\arguments{
\item{discdat}{dataframe; The genetic-geographic data by deme (K)}

\item{tempstart_params}{named double vector; vector of start parameters to serve as template.}

\item{momentum}{double; gamma parameter for momentum for adaptive learning rate}

\item{discsteps}{integer; the number of "steps" in the main \code{DISCent}
algorithm" \link{deme_inbreeding_spcoef}}

\item{downsample}{integer; the subset, or the number of a smaller random set of values, that you want to search in the combination grid}

\item{fstartmin}{double; the minimum value for F inbreeding coefficient in the search grid}

\item{fstartmax}{double; the maximum value for F inbreeding coefficient in the search grid}

\item{fstartby}{integer; the number of values between fstartmin and fstartmax to evaluate}

\item{mstartmin}{double; the minimum value for M global migration rate in the search grid}

\item{mstartmax}{double; the maximum value for M global migration rate in the search grid}

\item{mstartby}{integer; the number of values between mstartmin and mstartmax to evaluate}

\item{flearnmin}{double; the minimum value for F learning rate in the search grid}

\item{flearnmax}{double; the maximum value for F learning rate in the search grid}

\item{flearnby}{integer; the number of values between flearnmin and flearnmax to evaluate}

\item{mlearnmin}{double; the minimum value for M learning rate in the search grid}

\item{mlearnmax}{double; the maximum value for M learning rate in the search grid}

\item{mlearnby}{integer; the number of values between mlearnmin and mlearnmax to evaluate}
}
\description{
Find Optimal Start Gradient Parameters using Grid Search
}
\details{
Using a discrete grid search, identify the best start parameters for
inbreeding F values and M global migration parameter as well as the best
learning rates. We assume that each deme should have the same F start value for computational tractability.

The final cost for each set of start parameters is given by the gradient descent cost after \emph{n} iterations of
the main \code{DISCent} algorithm" \link{deme_inbreeding_spcoef}.
The search grid is defined by the user specified min, max, and steps ("by") of the
aforementioned parameters.
}
