% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.DISCresult}
\alias{is.DISCresult}
\title{Check if DISCresult S3 Class}
\usage{
is.DISCresult(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
Logical value: TRUE if object is of class DISCresult, FALSE otherwise
}
\description{
Tests whether an object is of class DISCresult
}
