% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.DISCresult}
\alias{print.DISCresult}
\title{Print DISCresult S3 Class}
\usage{
\method{print}{DISCresult}(x, ...)
}
\arguments{
\item{x}{Object to test}

\item{...}{Further arguments passed to or from other methods}
}
\value{
Invisibly returns the input object. Called for side effect of printing
}
\description{
S3 method for printing DISCresult objects with summary information
}
