% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_simulator.R
\name{run_forward_disc}
\alias{run_forward_disc}
\title{DISCent Forward Simulator}
\usage{
run_forward_disc(
  true_params,
  geodist_matrix,
  samples_per_deme,
  overdispersion = 200
)
}
\arguments{
\item{true_params}{named vector; True F_i and M values to simulate. At least one element must be named "m".}

\item{geodist_matrix}{named matrix of geodistances; must be a nxn geodistance matrix, with n corresponding to number of demes and deme names matching the true_param names}

\item{samples_per_deme}{integer vector; number of pairs within each deme}

\item{overdispersion}{numeric; overdispersion parameter in the beta-binomial model; Default: 200}
}
\value{
A dataframe of class "DISCsim" containing columns:
\itemize{
\item \code{Final_Fis}: Final inbreeding coefficient estimates
\item \code{Final_m}: Final migration rate estimate
\item \code{deme_key}: Mapping of deme names to indices
\item \code{cost}: Final cost function value(s)
}
}
\description{
Uses the DISCent equation, \eqn{r_ij = (f_i + f_j/2)e^(-d_{ij}/m)} to simulate
data expected genetic distances between sample pairs within demes. We assume a beta-binomial
model to create realistic "noise" amongst the pairs, where \eqn{Y|p ~ binom(n,p)} such that
\eqn{p ~ beta(\mu * \phi, (1 - \mu) * \phi)}, where \eqn{\mu} is average relatedness (defined by \eqn{\rij}) and \eqn{\phi}
is a concentration parameter of that relatedness. In our instance the beta-binomial returns counts (ie number of related segments),
we fix the segment denominator at 100. Overdispersion can be adjusted to produce more or less correlation
of relatedness
}
