% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.DISCresult}
\alias{summary.DISCresult}
\title{Summary of DISCresult S3 Class}
\usage{
\method{summary}{DISCresult}(object, ...)
}
\arguments{
\item{object}{DISCresult object from \code{\link{disc}} function}

\item{...}{Further arguments passed to or from other methods}
}
\value{
A list containing tidied final inbreeding coefficients and migration rate
}
\description{
S3 method for summarizing DISCresult objects
}
