% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tidyout.DISCresult}
\alias{tidyout.DISCresult}
\title{Tidy DISCresult Output}
\usage{
\method{tidyout}{DISCresult}(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
A list with two elements:
\itemize{
\item \code{Final_Fis}: Dataframe with deme names and final inbreeding coefficients
\item \code{Final_M}: Final migration rate estimate
}
}
\description{
S3 method for tidying DISCresult objects into user-friendly format
}
